Cleanup (#68)
* removed a duplicated helper funtion and moved it to helpers with a better name

* removed unused code and files, fixed a small isue in the encoder that would have resulted in a crash

* fixed signature of tryCastToNumber

* added scraps file and two functions we want to keep in it

* remoed test files

* added comment about funtion psotioning
