+++ mindsdb/libs/data_sources/__init__.py
@@ -1,5 +1,3 @@
-
-
 from mindsdb.libs.data_sources.array_to_cols_ds import ArrayToColsDS
 from mindsdb.libs.data_sources.file_ds import FileDS
 from mindsdb.libs.data_sources.window_ds import WindowDS
