+++ mindsdb/libs/helpers/general_helpers.py
